backend/requirements.txt
flask>=3.0
flask-cors>=5.0
flask-socketio>=5.4
python-dotenv>=1.0
sqlalchemy>=2.0
openai>=1.50
PyMuPDF>=1.24
charset-normalizer>=3.4
zep-cloud>=1.0
python-socketio>=5.11
eventlet>=0.38
